bugfix/trait_way_enum_wrong_serialize: move reusable Serializer to obp-commons, make trait way enum be sub type of SimpleEnum, Delete TransactionCommons type, it cause many wrong result when do serialize and deserialize.
