+++ obp-commons/pom.xml
@@ -27,6 +27,10 @@
             <artifactId>lift-common_${scala.version}</artifactId>
         </dependency>
         <dependency>
+            <groupId>net.liftweb</groupId>
+            <artifactId>lift-util_${scala.version}</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.scala-lang</groupId>
             <artifactId>scala-reflect</artifactId>
             <scope>compile</scope>
