+++ pom.xml
@@ -87,6 +87,11 @@
         <version>${lift.version}</version>
       </dependency>
       <dependency>
+        <groupId>net.liftweb</groupId>
+        <artifactId>lift-util_${scala.version}</artifactId>
+        <version>${lift.version}</version>
+      </dependency>
+      <dependency>
         <groupId>org.apache.commons</groupId>
         <artifactId>commons-collections4</artifactId>
         <version>4.4</version>
